# docs/swagger.yaml
definitions:
  middle.ResCode:
    enum:
    - 200
    - 201
    - 202
    - 203
    - 204
    - 205
    - 206
    - 207
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeInvalidParam
    - CodeUserExist
    - CodeUserNotExist
    - CodeInvalidPassword
    - CodeServerBusy
    - CodeNeedLogin
    - CodeInvalidToken
  middle.ResponseData:
    properties:
      code:
        $ref: '#/definitions/middle.ResCode'
      data: {}
      msg: {}
    type: object
  model.AddMenuAuthorityInput:
    properties:
      authorityId:
        description: 角色ID
        type: integer
      menus:
        items:
          $ref: '#/definitions/model.SysBaseMenu'
        type: array
    required:
    - authorityId
    type: object
...
  model.Empty:
    type: object
  model.Meta:
    properties:
      activeName:
        type: string
      closeTab:
        description: 自动关闭tab
        type: boolean
      icon:
        description: 菜单图标
        type: string
      keepAlive:
        description: 是否缓存
        type: boolean
      title:
        description: 菜单名
        type: string
    type: object
...
  model.SysAuthority:
    properties:
      authorityId:
        description: 角色ID
        type: integer
      authorityName:
        description: 角色名
        type: string
      children:
        items:
          $ref: '#/definitions/model.SysAuthority'
        type: array
      created_at:
        type: string
      dataAuthorityId:
        items:
          $ref: '#/definitions/model.SysAuthority'
        type: array
      defaultRouter:
        description: 默认菜单(默认dashboard)
        type: string
      menus:
        items:
          $ref: '#/definitions/model.SysBaseMenu'
        type: array
      parentId:
        description: 父角色ID
        type: integer
      updated_at:
        type: string
    type: object
  model.SysBaseMenu:
    properties:
      authoritys:
        items:
          $ref: '#/definitions/model.SysAuthority'
        type: array
      children:
        items:
          $ref: '#/definitions/model.SysBaseMenu'
        type: array
      created_at:
        type: string
      disabled:
        description: 是否在列表隐藏
        type: boolean
      hidden:
        description: 是否在列表隐藏
        type: boolean
      id:
        type: integer
      meta:
        allOf:
        - $ref: '#/definitions/model.Meta'
        description: 附加属性
      name:
        description: 路由name
        type: string
      parentId:
        description: 父菜单ID
        type: string
      path:
        description: 路由path
        type: string
      sort:
        description: 排序标记
        type: integer
      updated_at:
        type: string
    type: object
  model.SysBaseMenusResponse:
    properties:
      menus:
        items:
          $ref: '#/definitions/model.SysBaseMenu'
        type: array
    type: object
...
  service.HttpPath:
    properties:
      path:
        type: string
      pathType:
        $ref: '#/definitions/v1.PathType'
      serviceName:
        type: string
      servicePort:
        type: integer
    type: object
  service.IngressCreate:
    properties:
      hosts:
        additionalProperties:
          items:
            $ref: '#/definitions/service.HttpPath'
          type: array
        type: object
      label:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
    type: object
...
  v1.PathType:
    enum:
    - Exact
    - Prefix
    - ImplementationSpecific
    type: string
    x-enum-varnames:
    - PathTypeExact
    - PathTypePrefix
    - PathTypeImplementationSpecific
info:
  contact: {}
paths:
...
  /api/k8s/ingress/create:
    post:
      consumes:
      - application/json
      description: 创建ingress
      operationId: /api/k8s/ingress/create
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.IngressCreate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "新增成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 创建ingress
      tags:
      - ingress
  /api/k8s/ingress/del:
    delete:
      consumes:
      - application/json
      description: 删除ingress
      operationId: /api/k8s/ingress/del
      parameters:
      - description: ingress名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "删除成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 删除ingress
      tags:
      - ingress
  /api/k8s/ingress/detail:
    get:
      consumes:
      - application/json
      description: 获取ingress详情
      operationId: /api/k8s/ingress/detail
      parameters:
      - description: ingress名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data":""  }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 获取ingress详情
      tags:
      - ingress
  /api/k8s/ingress/list:
    get:
      consumes:
      - application/json
      description: 查看ingress列表
      operationId: /api/k8s/ingress/list
      parameters:
      - description: 过滤
        in: query
        name: filter_name
        type: string
      - description: 命名空间
        in: query
        name: namespace
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 分页限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data":""  }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 查看ingress列表
      tags:
      - ingress
...
  /api/k8s/ingress/update:
    put:
      consumes:
      - application/json
      description: 更新ingress
      operationId: /api/k8s/ingress/update
      parameters:
      - description: ingress名称
        in: query
        name: name
        required: true
        type: string
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: 更新内容
        in: query
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "更新成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 更新ingress
      tags:
      - ingress
  /api/k8s/namespace/create:
    put:
      consumes:
      - application/json
      description: 创建namespace
      operationId: /api/k8s/namespace/create
      parameters:
      - description: namespace名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "创建成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 创建namespace
      tags:
      - NameSpace
  /api/k8s/namespace/del:
    delete:
      consumes:
      - application/json
      description: 删除namespace
      operationId: /api/k8s/namespace/del
      parameters:
      - description: namespace名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": "删除成功}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 删除namespace
      tags:
      - NameSpace
  /api/k8s/namespace/detail:
    get:
      consumes:
      - application/json
      description: 获取NameSpace详情
      operationId: /api/k8s/namespace/detail
      parameters:
      - description: namespace名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data":data }'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 获取NameSpace详情
      tags:
      - NameSpace
  /api/k8s/namespace/list:
    get:
      consumes:
      - application/json
      description: 获取NameSpace列表
      operationId: /api/k8s/namespace/list
      parameters:
      - description: 过滤
        in: query
        name: filter_name
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 分页限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 200, msg="","data": service.NameSpaceResp}'
          schema:
            $ref: '#/definitions/middle.ResponseData'
      summary: 获取NameSpace列表
      tags:
      - NameSpace
...
  /api/menu/:authID/getMenuByAuthID:
    get:
      parameters:
      - description: 空
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Empty'
      produces:
      - application/json
      responses:
        "200":
          description: 获取用户动态路由,返回包括系统菜单详情列表
          schema:
            allOf:
            - $ref: '#/definitions/middle.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.SysBaseMenusResponse'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户动态路由
      tags:
      - AuthorityMenu
  /api/menu/add_base_menu:
    post:
      consumes:
      - application/json
      parameters:
      - description: 路由path, 父菜单ID, 路由name, 对应前端文件路径, 排序标记
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.SysBaseMenu'
      produces:
      - application/json
      responses:
        "200":
          description: 新增菜单
          schema:
            allOf:
            - $ref: '#/definitions/middle.ResponseData'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 新增菜单
      tags:
      - Menu
  /api/menu/add_menu_authority:
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AddMenuAuthorityInput'
      produces:
      - application/json
      responses:
        "200":
          description: 增加menu和角色关联关系
          schema:
            allOf:
            - $ref: '#/definitions/middle.ResponseData'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 增加menu和角色关联关系
      tags:
      - AuthorityMenu
  /api/menu/getBaseMenuTree:
    get:
      parameters:
      - description: 空
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Empty'
      produces:
      - application/json
      responses:
        "200":
          description: 获取系统菜单详情列表
          schema:
            allOf:
            - $ref: '#/definitions/middle.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.SysBaseMenusResponse'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户动态路由
      tags:
      - AuthorityMenu
...
swagger: "2.0"